11.November_ai-coding-assistant/requirements.txt
Flask==3.1.1
Flask-WTF==1.2.2
Werkzeug==3.1.3
google-auth==2.38.0
google-genai==1.24.0
WTForms==3.2.1
PyPDF2==3.0.1
python-docx==1.2.0